Allow taxon payload to handle non-'en' locales
In preparation to allow the Brexit taxon to be published with a 'cy' locale, we
need to ensure that the taxon payload that is built to send to the Publishing
API can handle non-'en' locales, as currently 'en' is the hardcoded locale
value. 'en' is provided as a default locale as all other taxons only need to be
published in that locale currently, however allowing an optional locale to be
provided will enable taxons to be published with a different locale if
necessary in future.